Bloglist - Autenticación de token
Funciona todo correctamente

+++ parte4/bloglist/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const logger = require('../utils/logger')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await 
@@ -24,12 +26,16 @@ blogsRouter.get('/:id', (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
-    const user = await User.findById(body.userId)
+    
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    
+    const user = await User.findById(decodedToken.id)
 
-    if (!user) {
-        // El usuario no fue encontrado en la base de datos, puedes enviar un mensaje de error o manejarlo de acuerdo a tus necesidades.
-        return response.status(404).json({ error: 'Usuario no encontrado' });
+    if (!body.likes){
+        body.like = 0
     }
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -39,37 +45,71 @@ blogsRouter.post('/', async (request, response, next) => {
     }
     )
 
+    try {
     const savedBlog = await blog.save()
-
+    requestLogger.info(`added ${blog.title} to the blog list`)    
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
-
+    logger.info(`blog linked to user ${user.username}`)
     response.json(savedBlog)
-})
-
-blogsRouter.delete('/:id', (request, response, next) => {
-    Blog.findByIdAndRemove(request.params.id)
-    .then(() => {
-        response.status(204).end()
-    })
-    .catch(error => next(error))
-})
-
-blogsRouter.put('/:id', (request, response, next) => {
-    const body = request.body 
-
-    const blog = {
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes
+    } catch (exception) {
+        next(exception)
     }
 
-    Blog.findByIdAndUpdate(request.params.id, blog, { new: true})
-    .then(updateBlog => {
-        response.json(updateBlog)
-    })
-    .catch(error => next(error))
+})
+
+blogsRouter.delete('/:id', async (request, response, next) => {
+    
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    
+    const user = await User.findById(decodedToken.id)
+
+    const blogToDelete = await Blog.findById(request.params.id)
+    
+    if ( blogToDelete.user._id.toString() === user._id.toString()){
+        try {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()  
+        } catch (exception) {
+            next(exception)
+        }
+    } else {
+        return response.status(401).json({error: `Unauthorized`})
+    }
+})
+
+blogsRouter.put('/:id', async (request, response, next) => {
+    const body = request.body
+
+    if (!body.likes) {
+        body.likes = 0
+    }
+    
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+
+    const blogToUpdate = await Blog.findById(request.params.id)
+
+    if ( blogToUpdate.user._id.toString() === user._id.toString() ) {
+        const blog = {
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes,
+        }
+
+        try {
+            const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+            logger.info(`blog ${blog.title} successfully updated`)
+            response.json(updatedBlog.toJSON())
+        } catch (exception) {
+            next(exception)
+        }
+    } else {
+        return response.status(401).json({ error: `Unauthorized` })
+    }
 })
 
 module.exports = blogsRouter
